refactor(builder): refactor validation logic into builder and rework structure
