+++ bundles/org.dataflowanalysis.analysis.dfd/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Require-Bundle: org.dataflowanalysis.analysis;bundle-version="1.0.0",
  org.dataflowanalysis.dfd.dataflowdiagram;bundle-version="0.1.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Automatic-Module-Name: org.dataflowanalysis.analysis.dfd
-Export-Package: org.dataflowanalysis.analysis.dfd.builder,
+Export-Package: org.dataflowanalysis.analysis.dfd,
  org.dataflowanalysis.analysis.dfd.core,
- org.dataflowanalysis.analysis.dfd,
  org.dataflowanalysis.analysis.dfd.resource
