+++ bundles/org.dataflowanalysis.analysis/META-INF/MANIFEST.MF
@@ -17,7 +17,6 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.24.0",
  org.dataflowanalysis.dfd.dataflowdiagram;bundle-version="0.1.0"
 Export-Package: 
  org.dataflowanalysis.analysis,
- org.dataflowanalysis.analysis.builder,
  org.dataflowanalysis.analysis.core,
  org.dataflowanalysis.analysis.resource,
  org.dataflowanalysis.analysis.utils
